- Completed rewrite for anchoring - Added a wedge for building pie charts. - Added examples.

+++ src/org/meta_environment/rascal/library/experiments/VL/Ellipse.java
@@ -18,7 +18,7 @@ public class Ellipse extends VELEM {
 	}
 
 	@Override
-	void bbox(int left, int top){
+	void bbox(float left, float top){
 		this.left = left;
 		this.top = top;
 		int lw = getLineWidthProperty();
