+++ src/org/meta_environment/rascal/library/experiments/VL/Shape.java
@@ -14,7 +14,7 @@ public class Shape extends Compose {
 	}
 	
 	@Override
-	void bbox(int left, int top){
+	void bbox(float left, float top){
 		this.left = left;
 		this.top = top;
 		width = 0;
